Move code inside textactivity
