+++ app/src/main/java/org/openobservatory/ooniprobe/activity/TextActivity.java
@@ -10,25 +10,45 @@ import android.view.MenuItem;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import org.apache.commons.io.FileUtils;
 import org.openobservatory.ooniprobe.R;
+import org.openobservatory.ooniprobe.client.callback.GetMeasurementJsonCallback;
+import org.openobservatory.ooniprobe.client.callback.GetMeasurementsCallback;
+import org.openobservatory.ooniprobe.common.ReachabilityManager;
+import org.openobservatory.ooniprobe.model.api.ApiMeasurement;
+import org.openobservatory.ooniprobe.model.database.Measurement;
 
 import androidx.annotation.Nullable;
+
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonParser;
+
+import java.io.File;
+import java.nio.charset.Charset;
+
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import localhost.toolkit.app.fragment.MessageDialogFragment;
+import okhttp3.Request;
 
 public class TextActivity extends AbstractActivity {
-	private static final String TEXT = "text";
+	private Measurement measurement;
+	private static final int TYPE_LOG = 1;
+	private static final int TYPE_JSON = 2;
+	private static final String TEST = "test";
+	private static final String TYPE = "type";
 	@BindView(R.id.textView) TextView textView;
 
-	public static Intent newIntent(Context context, String text) {
-		return new Intent(context, TextActivity.class).putExtra(TEXT, text);
+	public static Intent newIntent(Context context, int type, Measurement measurement) {
+		return new Intent(context, TextActivity.class).putExtra(TYPE, type).putExtra(TEST, measurement);
 	}
 
 	@Override protected void onCreate(@Nullable Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.text);
 		ButterKnife.bind(this);
-		textView.setText(getIntent().getStringExtra(TEXT));
+		measurement = (Measurement) getIntent().getSerializableExtra(TEST);
+		new Thread(() -> runOnUiThread(() -> showText())).start();
 	}
 
 	@Override public boolean onCreateOptionsMenu(Menu menu) {
@@ -44,6 +64,63 @@ public class TextActivity extends AbstractActivity {
 				return true;
 			default:
 				return super.onOptionsItemSelected(item);
+		}
+	}
+
+	public void showText(){
+		switch (getIntent().getIntExtra(TYPE, 0)) {
+			case TYPE_LOG:
+				//Try to open file, if it doesn't exist dont show Error dialog immediately but try to download the json from internet
+				try {
+					File entryFile = Measurement.getEntryFile(this, measurement.id, measurement.test_name);
+					String json = FileUtils.readFileToString(entryFile, Charset.forName("UTF-8"));
+					json = new GsonBuilder().setPrettyPrinting().disableHtmlEscaping().create().toJson(new JsonParser().parse(json));
+					textView.setText(json);
+				} catch (Exception e) {
+					e.printStackTrace();
+					if (ReachabilityManager.getNetworkType(this).equals(ReachabilityManager.NO_INTERNET)) {
+						new MessageDialogFragment.Builder()
+								.withTitle(getString(R.string.Modal_Error))
+								.withMessage(getString(R.string.Modal_Error_RawDataNoInternet))
+								.build().show(getSupportFragmentManager(), null);
+					}
+					getApiClient().getMeasurement(measurement.report_id, null).enqueue(new GetMeasurementsCallback() {
+						@Override
+						public void onSuccess(ApiMeasurement.Result result) {
+							getOkHttpClient().newCall(new Request.Builder().url(result.measurement_url).build()).enqueue(new GetMeasurementJsonCallback() {
+								@Override
+								public void onSuccess(String json) {
+									textView.setText(json);
+								}
+								@Override
+								public void onError(String msg) {
+									new MessageDialogFragment.Builder()
+											.withTitle(getString(R.string.Modal_Error))
+											.withMessage(msg)
+											.build().show(getSupportFragmentManager(), null);
+								}
+							});
+						}
+						@Override
+						public void onError(String msg) {
+							new MessageDialogFragment.Builder()
+									.withTitle(getString(R.string.Modal_Error))
+									.withMessage(msg)
+									.build().show(getSupportFragmentManager(), null);
+						}
+					});
+				}
+			case TYPE_JSON:
+				try {
+					File logFile = Measurement.getLogFile(this, measurement.result.id, measurement.test_name);
+					String log = FileUtils.readFileToString(logFile, Charset.forName("UTF-8"));
+					textView.setText(log);
+				} catch (Exception e) {
+					e.printStackTrace();
+					new MessageDialogFragment.Builder()
+							.withTitle(getString(R.string.Modal_Error_LogNotFound))
+							.build().show(getSupportFragmentManager(), null);
+				}
 		}
 	}
 }
